backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
APScheduler==3.10.4
requests==2.31.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.23
gunicorn==21.2.0
cryptography==41.0.7

# HTTP requests (for Taobao API)
urllib3==2.0.7

# Image processing
Pillow==10.4.0

# AI/ML
openai==1.12.0
